Tidy up and improve note message
