+++ XML_Adapter/XML_Adapter.csproj
@@ -144,6 +144,7 @@
     <Compile Include="CRUD\Create.cs" />
     <Compile Include="CRUD\CSProject\CreateCSProject.cs" />
     <Compile Include="CRUD\CSProject\ReadCSProject.cs" />
+    <Compile Include="CRUD\Default\ReadDefault.cs" />
     <Compile Include="CRUD\EnergyPlus\ReadEnergyPlus.cs" />
     <Compile Include="CRUD\GBXML\CreateGBXML.cs" />
     <Compile Include="CRUD\GBXML\ReadGBXML.cs" />
